+++ CHANGELOG.md
@@ -6,3 +6,7 @@
 - temporary root files (#8)
 - add initial core website, repo configuration and docs (#1)
 - initial commit
+
+### Fix
+
+- sync project json files (#15)